+++ pydidas/widgets/controllers/__init__.py
@@ -29,6 +29,7 @@ __all__ = []
 
 # import __all__ items from modules:
 from .manually_set_beamcenter_controller import *
+from .manually_set_integration_roi_controller import *
 
 # add modules' __all__ items to package's __all__ items and unclutter the
 # namespace by deleting the module references:
@@ -36,3 +37,8 @@ from . import manually_set_beamcenter_controller
 
 __all__.extend(manually_set_beamcenter_controller.__all__)
 del manually_set_beamcenter_controller
+
+from . import manually_set_integration_roi_controller
+
+__all__.extend(manually_set_integration_roi_controller.__all__)
+del manually_set_integration_roi_controller